+++ src/components/InputForm/Expiration/CardExpirationField.tsx
@@ -1,4 +1,3 @@
-import { useMemo, useRef } from "react";
 import NumberInput from "../../@common/NumberInput/NumberInput";
 import {
   NumberInputField,
@@ -6,8 +5,8 @@ import {
   NumberInputContainer,
   ErrorText,
 } from "../styles/CardField.styles";
-import { validateMonth, validateYear, validateExpiration } from "../validation";
 import { CardInfo } from "../../../hooks/useCardInfo";
+import useCardExpirationField from "./hooks/useCardExpirationField";
 
 interface CardExpirationFieldProps {
   cardInfo: CardInfo;
@@ -23,40 +22,8 @@ function CardExpirationField({
   handleCardInfo,
   maxLength,
 }: CardExpirationFieldProps) {
-  const yearInputRef = useRef<HTMLInputElement>(null);
-
-  const monthValidation = useMemo(
-    () => validateMonth(cardInfo.month, maxLength),
-    [cardInfo.month, maxLength]
-  );
-
-  const yearValidation = useMemo(
-    () => validateYear(cardInfo.year, maxLength),
-    [cardInfo.year, maxLength]
-  );
-
-  const expirationValidation = useMemo(
-    () => validateExpiration(cardInfo.month, cardInfo.year, maxLength),
-    [cardInfo.month, cardInfo.year, maxLength]
-  );
-
-  const errorMessage = useMemo(() => {
-    if (!monthValidation.isValid) return monthValidation.errorMessage;
-    if (!yearValidation.isValid) return yearValidation.errorMessage;
-    if (!expirationValidation.isValid) return expirationValidation.errorMessage;
-    return "";
-  }, [monthValidation, yearValidation, expirationValidation]);
-
-  const handleMonthChange = (value: string) => {
-    handleCardInfo("month", value);
-
-    if (value.length === maxLength) {
-      const monthNum = parseInt(value, 10);
-      if (monthNum >= 1 && monthNum <= 12) {
-        yearInputRef.current?.focus();
-      }
-    }
-  };
+  const { yearInputRef, validation, errorMessage, handleMonthChange } =
+    useCardExpirationField(cardInfo, maxLength, handleCardInfo);
 
   return (
     <NumberInputField>
@@ -70,7 +37,7 @@ function CardExpirationField({
           setValue={handleMonthChange}
           maxLength={maxLength}
           placeholder="MM"
-          isError={!monthValidation.isValid || !expirationValidation.isValid}
+          isError={!validation.month.isValid || !validation.expiration.isValid}
         />
         <NumberInput
           ref={yearInputRef}
@@ -80,7 +47,7 @@ function CardExpirationField({
           }}
           maxLength={maxLength}
           placeholder="YY"
-          isError={!yearValidation.isValid || !expirationValidation.isValid}
+          isError={!validation.year.isValid || !validation.expiration.isValid}
         />
       </NumberInputContainer>
       <ErrorText>{errorMessage}</ErrorText>
